feat: add registering firebase auth
